refactor: RouteClient 응답 분기를 객체로 캡슐화
